+++ tests/testthat/test_phasograms.R
@@ -1,5 +1,3 @@
-context("calcPhasogram")
-
 test_that("calcPhasogram(), estimateNRL() and calcAndCountDist() work properly", {
     ## create phasograms using MNase data from chr10:20000000-21000000
     bamf <- system.file("extdata", "phasograms", "mnase_mm10.bam", package = "swissknife")
@@ -17,13 +15,13 @@ test_that("calcPhasogram(), estimateNRL() and calcAndCountDist() work properly",
     ## check if the plotting function runs
     tf <- tempfile(fileext = ".pdf")
     pdf(file = tf)
-    expect_true(is.list(plotPhasogram(x = pg1)))
-    expect_true(is.list(plotPhasogram(x = pg1, verbosePlot = TRUE)))
+    expect_type(plotPhasogram(x = pg1), "list")
+    expect_type(plotPhasogram(x = pg1, verbosePlot = TRUE), "list")
     dev.off()
     unlink(tf)
 
     ## check expected results
-    expect_true(is.numeric(pg1))
+    expect_type(pg1, "double")
     expect_length(pg1, 3000L)
     expect_true(all(pg2 > pg1))
     expect_equal(pg1, pg3)
@@ -31,7 +29,7 @@ test_that("calcPhasogram(), estimateNRL() and calcAndCountDist() work properly",
     expect_true(sum(pg5) == 0L)
 
     nrl <- estimateNRL(pg1, usePeaks = 1:4)
-    expect_true(is.list(nrl))
+    expect_type(nrl, "list")
     expect_equal(nrl$nrl, 186.0)
     expect_length(nrl$nrl.CI95, 2L)